// src/common/types/redisUserAsStored.ts
type RedisUserAsStored = {
    name: string;
    avatarUrl: string;
    created: string;
    admin: string;
    tester: string;
    purged: string;
    booster: string;
    lives: string;
    erasers: string;
    superSpins: string;
    coins: string;
    seasonXp: string;
    chatBan: string;
    gameBan: string;
    appBan: string;
}

export default RedisUserAsStored;